// lib/vessel/src/doctypes/three-id/three-id-shape.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type PublicKeySecp256K1 = string;
export type PublicKeyX25519 = string;

/**
 * JSON Schema for 3id file
 */
export interface ThreeIdShape {
  doctype: string;
  owners: PublicKeySecp256K1[];
  content: {
    publicKeys: {
      encryption: PublicKeyX25519;
      signing: PublicKeySecp256K1;
    };
  };
}